+++ cotede/qctests/__init__.py
@@ -1,20 +1,20 @@
 #!/usr/bin/env python
 
-from qctests import *
-from global_range import global_range
-from descentPrate import descentPrate
-from location_at_sea import location_at_sea
-from density_inversion import density_inversion
-from profile_envelop import profile_envelop
-from woa_normbias import woa_normbias
-from anomaly_detection import anomaly_detection
-from possible_speed import possible_speed
-from rate_of_change import rate_of_change
-from cum_rate_of_change import cum_rate_of_change
-from morello2014 import morello2014
-from fuzzylogic import fuzzylogic
+from .qctests import *
+from .global_range import global_range
+from .descentPrate import descentPrate
+from .location_at_sea import location_at_sea
+from .density_inversion import density_inversion
+from .profile_envelop import profile_envelop
+from .woa_normbias import woa_normbias
+from .anomaly_detection import anomaly_detection
+from .possible_speed import possible_speed
+from .rate_of_change import rate_of_change
+from .cum_rate_of_change import cum_rate_of_change
+from .morello2014 import morello2014
+from .fuzzylogic import fuzzylogic
 
-from gradient import Gradient, gradient
-from spike import Spike, spike
-from bin_spike import Bin_Spike, bin_spike
-from tukey53H import Tukey53H, tukey53H_norm
+from .gradient import Gradient, gradient
+from .spike import Spike, spike
+from .bin_spike import Bin_Spike, bin_spike
+from .tukey53H import Tukey53H, tukey53H_norm