+++ cotede/qctests/fuzzylogic.py
@@ -21,7 +21,7 @@ def fuzzylogic(features, cfg):
     #   when it is missing some of the features? Sounds like the best guess
     #   possible.
     if not np.all([f in features for f in cfg['features']]):
-        print("Not all features (%s) required by morello2014 are available" %
+        print("Not all features (%s) required to fuzzyfy are available" %
                 cfg['features'].keys())
         try:
             return np.zeros(features[features.keys()[0]].shape, dtype='i1')
@@ -33,10 +33,12 @@ def fuzzylogic(features, cfg):
 
     # FIXME: As it is now, it will have no zero flag value. Think about cases
     #   where some values in a profile would not be estimated, hence flag=0
+    # I needed to use np.nonzeros because now uncertainty is a masked array,
+    #   to accept when a feature is masked.
     flags = np.zeros(features[cfg['features'].keys()[0]].shape, dtype='i1')
-    flags[uncertainty <= 0.29] = 1
-    flags[(uncertainty > 0.29)  & (uncertainty <= 0.34)] = 2
-    flags[(uncertainty > 0.34)  & (uncertainty <= 0.72)] = 3
-    flags[uncertainty > 0.72] = 4
+    flags[np.nonzero(uncertainty <= 0.29)] = 1
+    flags[np.nonzero((uncertainty > 0.29)  & (uncertainty <= 0.34))] = 2
+    flags[np.nonzero((uncertainty > 0.34)  & (uncertainty <= 0.72))] = 3
+    flags[np.nonzero(uncertainty > 0.72)] = 4
 
     return flags